Fix compilation error on macOS
This change fixes compilation error when using macOS 14.5 and
Apple clang version 15.0.0 (clang-1500.3.9.4).

It seems that there is an operator<< defined for the time point
which internally relies on std::format. The std::format, however
is only available with deployment target 13.3.

A simple fix is to check for equality operator to be true.