Refactor: Remove unsafe RepresentationView
Splitting representation storage from its access had some
downsides of needing extra care when objects are getting
copied or assigned to ensure validness of the view.

Combines storage with a flexible access to it.

The side effect of this change is use naming convention
which is more native to this project.