Fix tests failure on Intel x64 MSVC builds
There is nothing really wrong with the code, there might be
some difference in in the way how FMA is used by compilers,
or could be difference in the modulo from the STD.

Adjust the test code and raise the epsilon a bit. Now it uses
the same epsilon than in ERFA tests for the similar functions.